# internal/platform/httpserver/openapi.yaml
openapi: 3.0.3
info:
  title: Backend CRM Piposmart
  version: 0.15.4-sprint-14f-identity
  description: API CRM Piposmart dengan Partner Type, Partner (Encrypted & Masked Bank Account), PIC Assignment, Call Interactions, Lead Referrals, Subscription Order, Reconciliation Sprint 11, Partner Commission Sprint 12, penambahan route trash/unscoped owner-outlet dan rekap status langganan outlet Sprint 14b/14c, histori import persisten dengan viewer/download file asli Sprint 14 history, route /all dan /all-deleted lintas modul untuk kebutuhan frontend mengambil full dataset tanpa pagination, serta penambahan manajemen akun Admin dan Supervisor beserta reset password lintas role pada Sprint 14f.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/imports:
    post:
      tags: [Imports]
      summary: Upload Excel file untuk import
      operationId: uploadImport
      description: Hanya ADMIN. Dedup berdasarkan SHA-256 file â€” upload file yang sama mengembalikan batch yang sudah ada. Validasi berjalan async lewat worker (job_queue). Profil (OWNER_OUTLET/NON_REGISTER) bisa dideklarasikan eksplisit atau dibiarkan auto-detect dari header kolom.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                profile:
                  type: string
                  enum: [OWNER_OUTLET, NON_REGISTER]
      responses:
        "201":
          description: Batch berhasil diupload (atau batch existing dikembalikan bila file sama)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
        "400":
          description: File tidak valid (bukan .xlsx, terlalu besar, atau kosong)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Akses ditolak (bukan ADMIN)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    get:
      tags: [Imports]
      summary: List import batch
      operationId: listImportBatches
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: profile
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Daftar batch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /api/v1/imports/{id}:
    get:
      tags: [Imports]
      summary: Detail import batch
      operationId: getImportBatch
      description: Status dan counts (total/valid/invalid/committed) jadi preview sebelum commit.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Detail batch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
        "404":
          description: Batch tidak ditemukan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/v1/imports/{id}/file:
    get:
      tags: [Imports]
      summary: Buka file Excel asli dari batch import
      operationId: viewImportOriginalFile
      description: Hanya ADMIN. Mengembalikan workbook asli yang pernah diupload agar frontend dapat merender halaman viewer Excel sendiri.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: File workbook asli
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
        "404":
          description: Batch tidak ditemukan atau file asli sudah tidak tersedia
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/v1/imports/{id}/file/download:
    get:
      tags: [Imports]
      summary: Unduh file Excel asli dari batch import
      operationId: downloadImportOriginalFile
      description: Hanya ADMIN. Mengembalikan workbook asli sebagai attachment unduhan.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: File workbook asli sebagai attachment
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
        "404":
          description: Batch tidak ditemukan atau file asli sudah tidak tersedia
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/v1/imports/{id}/rows:
    get:
      tags: [Imports]
      summary: List baris dalam satu batch
      operationId: listImportRows
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: status
          in: query
          description: PENDING, VALID, INVALID, COMMITTED, atau COMMIT_FAILED
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Daftar baris
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /api/v1/imports/{id}/rejected-rows/export:
    get:
      tags: [Imports]
      summary: Unduh baris invalid sebagai CSV
      operationId: exportRejectedRows
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: File CSV berisi baris invalid + alasan
          content:
            text/csv:
              schema:
                type: string
                format: binary
  /api/v1/imports/{id}/commit:
    post:
      tags: [Imports]
      summary: Commit batch yang sudah divalidasi
      operationId: commitImportBatch
      description: Hanya ADMIN. Commit pertama hanya boleh saat batch berstatus VALIDATED. Retry aman untuk status COMMITTING atau COMMITTED. Berjalan async; poll GET /imports/{id} untuk status.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "202":
          description: Commit di-enqueue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
        "400":
          description: Batch belum VALIDATED atau sudah COMMITTED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Akses ditolak (bukan ADMIN)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /api/v1/imports/all:
    get:
      tags: [Imports]
      summary: List semua import batch tanpa pagination
      operationId: listImportBatchesAll
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: profile
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List semua batch import
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /api/v1/imports/all-deleted:
    get:
      tags: [Imports]
      summary: Alias full import batch untuk kompatibilitas frontend
      operationId: listImportBatchesAllDeleted
      description: Modul import batch tidak memakai soft delete. Route ini adalah alias kompatibilitas frontend.
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: profile
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List semua batch import
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /api/v1/imports/{id}/rows/all:
    get:
      tags: [Imports]
      summary: List semua row import tanpa pagination
      operationId: listImportRowsAll
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List semua row import
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /api/v1/imports/{id}/rows/all-deleted:
    get:
      tags: [Imports]
      summary: Alias full row import untuk kompatibilitas frontend
      operationId: listImportRowsAllDeleted
      description: Modul import row tidak memakai soft delete. Route ini adalah alias kompatibilitas frontend.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List semua row import
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    SuccessEnvelope:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          additionalProperties: true
        meta:
          type: object
          required: [request_id]
          properties:
            request_id:
              type: string
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              nullable: true
            request_id:
              type: string
...